# SecundaTask/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация по API",
        "description": "Документация REST API для поиска, по различным данным, организаций. Так же для вывода всех зданий, которые есть в базе.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Основной сервер API."
        }
    ],
    "paths": {
...
        "/api/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получение организации по ID",
                "operationId": "37c671d3f12ac513261f677f9b7711a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Организация найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Организация по введенному ID найдена!"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Organization"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не передан или неверный API ключ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Не передан API ключ"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Данные не найдены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Данные не найдены"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/activity": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организации по виду деятельности(Тут указываем либо id активности, либо название активности)",
                "operationId": "36d0047d953dd39eaeeb054b6e4893f0",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "query",
                        "description": "ID деятельности",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "activity_name",
                        "in": "query",
                        "description": "Название деятельности",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "include_children",
                        "in": "query",
                        "description": "Включать дочерние виды деятельности",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций по выбранной деятельности",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Организации по введенной активности найдены!"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не передан или неверный API ключ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Не передан API ключ"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Данные не найдены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Данные не найдены"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/organizations/name": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организации по названию",
                "operationId": "b597c72a522ce1a3f09405292bde3d30",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Название организации",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Организация найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "code": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Организация с введенным названием найдена!"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не передан или неверный API ключ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Не передан API ключ"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Данные не найдены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Данные не найдены"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Organization": {
                "title": "Организация",
                "description": "Модель организации с вложенными связями",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ПАО МоторЭлектроТранс"
                    },
                    "building_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "office_number": {
                        "type": "string",
                        "example": "81"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-20T23:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-20T23:00:00.000000Z"
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Phone"
                        }
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-key"
            }
        }
    },
...
    "security": [
        {
            "ApiKeyAuth": []
        }
    ]
}